Clean up C++ surface file structure
