+++ CMakeLists.txt
@@ -437,11 +437,13 @@ set(LIBOPENMC_FORTRAN_SRC
 set(LIBOPENMC_CXX_SRC
   src/error.h
   src/hdf5_interface.h
-  src/random_lcg.h
   src/random_lcg.cpp
-  src/surface_header.C
-  src/pugixml/pugixml.hpp
-  src/pugixml/pugixml.cpp)
+  src/random_lcg.h
+  src/surface.cpp
+  src/surface.h
+  src/xml_interface.h
+  src/pugixml/pugixml.cpp
+  src/pugixml/pugixml.hpp)
 add_library(libopenmc SHARED ${LIBOPENMC_FORTRAN_SRC} ${LIBOPENMC_CXX_SRC})
 set_target_properties(libopenmc PROPERTIES OUTPUT_NAME openmc)
 add_executable(${program} src/main.F90)
