+++ src/error.h
@@ -1,6 +1,7 @@
 #ifndef ERROR_H
 #define ERROR_H
 
+#include <cstring>
 #include <string>
